Change the token path in the precomputed clusters
Pilot precomputes clusters of CDS response based on the events
on k8s API server. The events on k8s API server do not contain
Envoy node metadata. When an Envoy sends a CDS request to Pilot,
Pilot sends precomputed CDS clusters to Envoy, which contains
the default SDS token path. The SDS token path in the precomputed
CDS needs to be changed to the SDS_TOKEN_PATH in the Envoy node
metadata.

clusters = append(clusters,
        sidecarScope.CDSOutboundClusters[util.NoProxyLocality]...)